refactor: useProducts 훅 분리
